botTradeLambda/aws_requirements.txt
alpaca-py==0.10.0
numpy==1.25.2
pytz==2023.3
torch==2.0.1